Added centred blocker to the virtual list.
