+++ loveship/components/layout/VirtualList.tsx
@@ -1,5 +1,5 @@
 import { withMods } from '@epam/uui-core';
-import { VirtualList as uuiVirtualList, VirtualListProps } from '@epam/uui-components';
+import { VirtualList as uuiVirtualList, VirtualListProps } from '@epam/uui';
 import css from './VirtualList.module.scss';
 
 export interface VirtualListMods {}
