+++ uui-core/src/hooks/useVirtualList/utils.ts
@@ -64,6 +64,7 @@ export const getUpdatedRowsInfo = (
         virtualListInfo.rowsCount,
         virtualListInfo.listOffset,
     );
+    console.log(virtualListInfo.rowsCount);
     return {
         estimatedHeight,
         rowHeights,
